+++ backend/migrations/versions/2025_12_20_1344_e00067cc5c0e_init.py
@@ -143,7 +143,7 @@ def upgrade() -> None:
         sa.Column("destination_latitude", sa.REAL(), nullable=False),
         sa.Column("destination_longitude", sa.REAL(), nullable=False),
         sa.Column("transport_type", sa.String(length=20), nullable=False),
-        sa.Column("distance_km", sa.REAL(), nullable=False),
+        sa.Column("distance_km", sa.SmallInteger(), nullable=False),
         sa.Column("traveled_on", sa.Date(), nullable=False),
         sa.Column("traveled_on_precision", sa.String(length=5), nullable=False),
         sa.PrimaryKeyConstraint("id"),
